Rename CoreUtils to NumberFormatUtils
